--- lib/src/main/java/org/braid/secret/society/voicevox4j/internal/structs/VoicevoxAccelerationMode.java
@@ -1,17 +1,0 @@
-package org.braid.secret.society.voicevox4j.internal.structs;
-
-import com.sun.jna.ptr.IntByReference;
-
-public class VoicevoxAccelerationMode extends IntByReference {
-  public static final int VOICEVOX_ACCELERATION_MODE_AUTO = 0;
-  public static final int VOICEVOX_ACCELERATION_MODE_CPU = 1;
-  public static final int VOICEVOX_ACCELERATION_MODE_GPU = 2;
-
-  public VoicevoxAccelerationMode() {
-    super();
-  }
-
-  public VoicevoxAccelerationMode(int value) {
-    super(value);
-  }
-}
