--- lib/src/main/java/org/braid/secret/society/voicevox4j/internal/structs/VoicevoxVoiceModelFile.java
@@ -1,15 +1,0 @@
-package org.braid.secret.society.voicevox4j.internal.structs;
-
-import com.sun.jna.Pointer;
-
-public class VoicevoxVoiceModelFile extends Pointer {
-
-  /**
-   * Create from native pointer.  Don't use this unless you know what you're doing.
-   *
-   * @param peer
-   */
-  public VoicevoxVoiceModelFile(long peer) {
-    super(peer);
-  }
-}
