+++ lib/src/test/java/org/braid/society/secret/voicevox4j/DepsTest.java
@@ -1,0 +1,14 @@
+package org.braid.society.secret.voicevox4j;
+
+import java.nio.file.Path;
+import org.junit.jupiter.api.BeforeEach;
+
+public class DepsTest {
+
+  private Voicevox voicevox;
+
+  @BeforeEach
+  void setUp() {
+    voicevox = new Voicevox(Path.of(""));
+  }
+}
